Consolidate and simplify package names
