+++ tests/packages/importlib_editable/pkg/CMakeLists.txt
@@ -4,5 +4,5 @@ install(TARGETS emod_a DESTINATION pkg/)
 file(WRITE "${CMAKE_CURRENT_BINARY_DIR}/testfile" "This is the file")
 install(FILES "${CMAKE_CURRENT_BINARY_DIR}/testfile" DESTINATION pkg/)
 
-add_subdirectory(subpkg1)
-add_subdirectory(subpkg2)
+add_subdirectory(sub_a)
+add_subdirectory(sub_b)
