+++ tests/packages/importlib_editable/pkg/sub_b/sub_c/emod_d.c
