+++ tests/packages/importlib_editable/pkg/sub_b/sub_d/emod_e.c
