--- tests/packages/importlib_editable/pkg/subpkg2/CMakeLists.txt
@@ -1,8 +1,0 @@
-python_add_library(emod_c MODULE emod_c.c WITH_SOABI)
-
-install(TARGETS emod_c DESTINATION pkg/subpkg2)
-file(WRITE "${CMAKE_CURRENT_BINARY_DIR}/testfile" "This is the file")
-install(FILES "${CMAKE_CURRENT_BINARY_DIR}/testfile" DESTINATION pkg/subpkg2)
-
-add_subdirectory(subsubpkg1)
-add_subdirectory(subsubpkg2)
